コンパイル時のwarningを解消

+++ kern/mm.c
@@ -141,7 +141,7 @@ void kfree(void *va)
 {
     struct page *page = page_find_by_address(va);
     if (page) {
-        if (va == 0xffff000000bfd000) {
+        if ((uint64_t)va == 0xffff000000bfd000) {
             debug("page->ref: %d", page->ref);
         }
         if (--page->ref <= 0)
